Refactor version related code to it's own class
